--- src/components/Header/Header.Container.js
@@ -1,9 +1,0 @@
-import { connect } from "react-redux";
-import * as sessionAction from "../../redux/session/sessionAction";
-import Header from "./Header";
-
-const mapDispatchToProps = {
-	onLogOut: sessionAction.logOut,
-};
-
-export default connect(null, mapDispatchToProps)(Header);
