+++ src/pages/HomePage/HomePage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 //import style from "./HomePage.module.css";
-import Header from '../../components/Header/Header.Container';
+import Header from '../../containers/Header';
 import TableTransactions from '../../components/TableTransactions/TableTransactions';
 
 const HomePage = () => {
