+++ src/pages/LoginPage/LoginPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 //import style from "./LoginPage.module.css";
 import EmptyLayout from '../../components/EmptyLayout/EmptyLayout';
-import Loginization from '../../components/Loginization/LoginizationContainer';
+import Loginization from '../../containers/Loginization';
 
 const LoginPage = () => {
   return (
