+++ src/redux/operations/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import sessionOperation from './app/sessionOperation';
+
+export default combineReducers({
+  sessionOperation,
+});
