+++ src/redux/reducers/app/index.js
@@ -1,3 +1,4 @@
 import { combineReducers } from 'redux';
+import { rootReducer } from '../app/sessionReducer';
 
-export default combineReducers({});
+export default combineReducers({ rootReducer });
