+++ src/redux/selectors/app/sessionSelectors.js
@@ -1,3 +1,3 @@
 export const getIsAuthenticated = (state) => {
-	return state.authenticated;
+  return state.authenticated;
 };
