+++ src/redux/type.js
@@ -1,0 +1,9 @@
+export const Type = {
+  LOGIN_REQUEST: 'LOGIN_REQUEST',
+  LOGIN_SUCCESS: 'LOGIN_SUCCESS',
+  LOGIN_ERROR: 'LOGIN_EERROR',
+  SIGNUP_REQUEST: 'SIGNUP_REQUEST',
+  SIGNUP_SUCCESS: 'SIGNUP_SUCCESS',
+  SIGNUP_ERROR: 'SIGNUP_EERROR',
+  LOGOUT: 'LOGOUT',
+};
